requirements.txt
# Core dependencies for BrainTransactionsManager
# Blessed by Goddess Laxmi for Infinite Abundance 🙏

# Data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0

# Configuration management
python-dotenv>=0.19.0
pydantic>=1.10.0

# Database connectivity
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Trading APIs
alpaca-trade-api>=3.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1
requests>=2.28.0

# Logging and monitoring
structlog>=22.0.0

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Optional: Telegram integration
# python-telegram-bot>=20.0

# Development tools (optional)
# black>=22.0.0
# flake8>=5.0.0
# mypy>=0.991
